updated vehicles search component to separate logic from presentation
